+++ java/code/src/com/redhat/rhn/domain/kickstart/KickstartFactory.java
@@ -179,8 +179,8 @@ public class KickstartFactory extends HibernateFactory {
     public static KickstartData lookupKickstartDataByIdAndOrg(Org orgIn, Long ksid) {
         return (KickstartData)  HibernateFactory.getSession()
                 .getNamedQuery("KickstartData.findByIdAndOrg")
-                .setLong("id", ksid.longValue())
-                .setLong("org_id", orgIn.getId().longValue())
+                .setLong("id", ksid)
+                .setLong("org_id", orgIn.getId())
                 .uniqueResult();
     }
 
@@ -212,7 +212,7 @@ public class KickstartFactory extends HibernateFactory {
         return (KickstartData) HibernateFactory.getSession().
                 getNamedQuery("KickstartData.findByLabelAndOrg")
                 .setString("label", label)
-                .setLong("org_id", orgId.longValue())
+                .setLong("org_id", orgId)
                 .uniqueResult();
     }
 
@@ -231,7 +231,7 @@ public class KickstartFactory extends HibernateFactory {
         return (KickstartData) HibernateFactory.getSession().
                 getNamedQuery("KickstartData.findByCILabelAndOrg")
                 .setString("label", label)
-                .setLong("org_id", orgId.longValue())
+                .setLong("org_id", orgId)
                 .uniqueResult();
     }
 
@@ -517,7 +517,7 @@ public class KickstartFactory extends HibernateFactory {
         if (org != null) {
             query = session.getNamedQuery("CryptoKey.findByDescAndOrg")
                            .setString("description", description)
-                           .setLong("org_id", org.getId().longValue());
+                           .setLong("org_id", org.getId());
         }
         else {
             query = session.getNamedQuery("CryptoKey.findByDescAndNullOrg")
@@ -537,7 +537,7 @@ public class KickstartFactory extends HibernateFactory {
         //look for Kickstart data by id
         session = HibernateFactory.getSession();
         retval = session.getNamedQuery("CryptoKey.findByOrg")
-                .setLong("org_id", org.getId().longValue())
+                .setLong("org_id", org.getId())
                 .list();
         return retval;
     }
@@ -553,7 +553,7 @@ public class KickstartFactory extends HibernateFactory {
         //look for Kickstart data by id
         session = HibernateFactory.getSession();
         retval = session.getNamedQuery("SslCryptoKey.findByOrg")
-                .setLong("org_id", org.getId().longValue())
+                .setLong("org_id", org.getId())
                 .list();
         return retval;
     }
@@ -570,8 +570,8 @@ public class KickstartFactory extends HibernateFactory {
         //look for Kickstart data by id
         session = HibernateFactory.getSession();
         retval = (CryptoKey) session.getNamedQuery("CryptoKey.findByIdAndOrg")
-                .setLong("key_id", keyId.longValue())
-                .setLong("org_id", org.getId().longValue())
+                .setLong("key_id", keyId)
+                .setLong("org_id", org.getId())
                 .uniqueResult();
         return retval;
     }
@@ -588,8 +588,8 @@ public class KickstartFactory extends HibernateFactory {
         //look for Kickstart data by id
         session = HibernateFactory.getSession();
         retval = (SslCryptoKey) session.getNamedQuery("SslCryptoKey.findByIdAndOrg")
-                .setLong("key_id", keyId.longValue())
-                .setLong("org_id", org.getId().longValue())
+                .setLong("key_id", keyId)
+                .setLong("org_id", org.getId())
                 .uniqueResult();
         return retval;
     }
@@ -623,7 +623,7 @@ public class KickstartFactory extends HibernateFactory {
         retval = (KickstartableTree)
                 session.getNamedQuery("KickstartableTree.findByLabelAndOrg")
                 .setString("label", label)
-                .setLong("org_id", org.getId().longValue())
+                .setLong("org_id", org.getId())
                 .uniqueResult();
         // If we don't find by label + org then
         // we try by label and NULL org (RHN owned channel)
@@ -669,8 +669,8 @@ public class KickstartFactory extends HibernateFactory {
         String query = "KickstartableTree.findByChannelAndOrg";
         session = HibernateFactory.getSession();
         retval = session.getNamedQuery(query).
-                setLong("channel_id", channelId.longValue()).
-                setLong("org_id", org.getId().longValue())
+                setLong("channel_id", channelId).
+                setLong("org_id", org.getId())
                 //Retrieve from cache if there
                 .setCacheable(true).list();
         return retval;
@@ -688,7 +688,7 @@ public class KickstartFactory extends HibernateFactory {
         String query = "KickstartableTree.findByChannelAndNullOrg";
         session = HibernateFactory.getSession();
         retval = session.getNamedQuery(query)
-                .setLong("channel_id", channelId.longValue())
+                .setLong("channel_id", channelId)
                 // Retrieve from cache if there
                 .setCacheable(true).list();
         return retval;
@@ -710,8 +710,8 @@ public class KickstartFactory extends HibernateFactory {
         query = "KickstartableTree.findByChannel";
         session = HibernateFactory.getSession();
         retval = session.getNamedQuery(query).
-                setLong("channel_id", channelId.longValue()).
-                setLong("org_id", org.getId().longValue()).
+                setLong("channel_id", channelId).
+                setLong("org_id", org.getId()).
                 list();
         return retval;
     }
@@ -777,8 +777,8 @@ public class KickstartFactory extends HibernateFactory {
         if (treeId != null && org != null) {
             session = HibernateFactory.getSession();
             Query query = session.getNamedQuery(queryName);
-            query.setLong("org_id", org.getId().longValue());
-            query.setLong("tree_id", treeId.longValue());
+            query.setLong("org_id", org.getId());
+            query.setLong("tree_id", treeId);
             //Retrieve from cache if there
             retval = (KickstartableTree)
                     query.setCacheable(true).uniqueResult();
@@ -801,7 +801,7 @@ public class KickstartFactory extends HibernateFactory {
         Session session = null;
         session = HibernateFactory.getSession();
         List ksessions = session.getNamedQuery("KickstartSession.findByServer")
-                .setLong("server", sidIn.longValue())
+                .setLong("server", sidIn)
                 .list();
         if (ksessions.size() > 0) {
             return (KickstartSession) ksessions.iterator().next();
@@ -859,7 +859,7 @@ public class KickstartFactory extends HibernateFactory {
     public static List lookupAllKickstartSessionsByServer(Long sidIn) {
         Session session = HibernateFactory.getSession();
         return session.getNamedQuery("KickstartSession.findByServer")
-                .setLong("server", sidIn.longValue())
+                .setLong("server", sidIn)
                 .list();
     }
 
@@ -895,9 +895,9 @@ public class KickstartFactory extends HibernateFactory {
             session = HibernateFactory.getSession();
             Query query = session.
                     getNamedQuery("KickstartableTree.verifyTreeAssignment");
-            query.setLong("channel_id", channelId.longValue());
-            query.setLong("org_id", orgId.longValue());
-            query.setLong("tree_id", treeId.longValue());
+            query.setLong("channel_id", channelId);
+            query.setLong("org_id", orgId);
+            query.setLong("tree_id", treeId);
             Object tree = query.uniqueResult();
             retval = (tree != null);
         }
@@ -979,7 +979,7 @@ public class KickstartFactory extends HibernateFactory {
         String query = "KickstartData.lookupByTreeId";
         Session session = HibernateFactory.getSession();
         return session.getNamedQuery(query)
-                .setLong("kstree_id", tree.getId().longValue())
+                .setLong("kstree_id", tree.getId())
                 .list();
     }
